feat: set session api

+++ src/dto/request.rs
@@ -7,7 +7,7 @@ pub struct RefreshRequest {
 #[derive(Debug, Clone, Default, Deserialize, Serialize)]
 pub struct SetSessionRequest {
     pub subscription_status: Option<bool>,
-    pub credits_remaining: Option<i64>,
+    pub credits_remaining: Option<f64>,
     pub preferences: Option<serde_json::Value>,
     pub session_metadata: Option<serde_json::Value>,
 }
