+++ src/entity/session.rs
@@ -2,7 +2,7 @@ use chrono::{DateTime, Utc};
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize, DeriveEntityModel)]
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize, DeriveEntityModel)]
 #[sea_orm(table_name = "sessions")]
 pub struct Model {
     #[sea_orm(primary_key)]
@@ -10,7 +10,7 @@ pub struct Model {
     #[sea_orm(unique, indexed)]
     pub user_id: i64,
     pub subscription_status: bool,
-    pub credits_remaining: i64,
+    pub credits_remaining: f64,
     pub last_active_timestamp: i64,
     pub preferences: serde_json::Value,
     pub session_metadata: serde_json::Value,
